+++ lib/potassium/assets/seeds/seeds.rb
@@ -1,0 +1,10 @@
+# This file should contain all the record creation needed to seed the database
+# with its default values.
+# The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
+# Write the code in a way that can be executed multiple times without duplicating the information.
+#
+# For example:
+#
+# Country.create(name: "Chile") # BAD
+# Country.find_or_create_by(name: "Chile") # GOOD
+#